Moved redis client
