+++ client/redis/redis_test.go
